requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
lxml>=5.0.0
deep-translator==1.11.4

# Web Framework
streamlit==1.31.0

# Scheduling
APScheduler==3.10.4

# Data Processing
pandas>=2.2.0

# Utilities
python-dotenv==1.0.0
python-dateutil>=2.8.0

# AI Providers
openai>=1.12.0
groq>=0.4.0

# Article Extraction (Playwright alternative for Streamlit Cloud)
newspaper3k==0.2.8
trafilatura>=1.6.0
ddgs>=0.9.0

# Browser Automation (Optional - Only for local development)
# playwright>=1.40.0
# NOTE: Playwright is NOT supported on Streamlit Cloud
# It requires browser binaries that cannot be installed on Streamlit Cloud
# Use newspaper3k/trafilatura for article extraction on cloud deployments